#727 To avoid cyclic dependencies buildFromParams and render must be in the same file as the class
